% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literal.R
\name{literal_classes}
\alias{literal_classes}
\alias{class_literal}
\alias{class_literal_union}
\title{Literal classes}
\usage{
class_literal(value)

class_literal_union(...)
}
\arguments{
\item{value}{A literal value to construct a class for.}

\item{...}{Literal values to construct a class union for.}
}
\description{
Literal classes represent objects whose class is determined by their literal
value. They are mostly useful in class unions, where they can model a finite
set of valid values.
}
\examples{
class_literal("GET")

class_http_method <- class_literal_union("GET", "POST", "PUT", "DELETE")
class_http_method

Request <- new_class("Request", properties = list(method = class_http_method))
Request
Request(method = "GET")
try(Request(method = "foo"))
}
