% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-register.R
\name{methods_check_compatibility}
\alias{methods_check_compatibility}
\title{Check whether two generics can be merged}
\usage{
methods_check_compatibility(genericA, genericB)
}
\description{
Currently, this is as conservative as possible and will not allow any
differences between methods provided by different sources.
}
\note{
It would be interesting to explore how lenient this could be.
\itemize{
\item Perhaps a difference of the dispatch body could individually mask the
dispatching body of the existing generic.
\item Perhaps mismatches in arguments (so long as the dispatch args are
consistent) need not be consistent
\item Perhaps differences in dispatch args can force the method to be used with
only positional arguments to disambiguate dispatch args.
\item Maybe these more lenient interop behaviors could be user-configurable
}
}
