% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-register.R
\name{methods_register}
\alias{methods_register}
\title{Register methods on generics from an environment}
\usage{
methods_register(libpath, package = packageName(parent.frame()))
}
\arguments{
\item{from}{An environment whose generics should be registered. By default,
the parent environment is used.}
}
\description{
Register methods on generics from an environment
}
