% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-register.R
\name{unregistrar}
\alias{unregistrar}
\title{Generate a function to remove a package's contributed methods}
\usage{
unregistrar(package, generic)
}
\arguments{
\item{package}{A package name, whose namespace may be the originating
namespace for some number of methods registered to generic}

\item{generic}{An R7 generic}
}
\value{
A function, which may be used as a package event hook to remove a
package's contributed methods
}
\description{
Generate a function to remove a package's contributed methods
}
\keyword{internal}
